+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 0.2.2
 * Change test lib to buster-node + referee
 * Set min node engine to >= v0.8.0
+* Add blink on failure flag to buildlight command #23
 
 ### 0.2.1
 * BuildLight notifier blinks red on build failure